requirements.txt
py-graphql-client
brotli
pydash
pika